[FLINK-29655] Move FlinkResourceListener to api
