+++ flink-kubernetes-operator/src/test/java/org/apache/flink/kubernetes/operator/listener/TestingListener.java
@@ -21,6 +21,7 @@ package org.apache.flink.kubernetes.operator.listener;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.kubernetes.operator.api.FlinkDeployment;
 import org.apache.flink.kubernetes.operator.api.FlinkSessionJob;
+import org.apache.flink.kubernetes.operator.api.listener.FlinkResourceListener;
 import org.apache.flink.kubernetes.operator.api.status.FlinkDeploymentStatus;
 import org.apache.flink.kubernetes.operator.api.status.FlinkSessionJobStatus;
